asr/train/train_parakeet_tdt_with_new_tokenizer/requirements.txt
# Training requirements for Parakeet TDT with multilingual tokenizer
# NeMo and dependencies
nemo_toolkit[asr]>=1.22.0
pytorch-lightning>=2.0.0
omegaconf>=2.3.0

# Model handling
torch>=2.1.0
transformers>=4.35.0
huggingface_hub>=0.19.0

# Tokenizer
sentencepiece>=0.1.99

# Training optimization
bitsandbytes>=0.41.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0

# Utilities
tensorboard>=2.14.0
